Update component to make CCM container image configurable
Also update manifest fetching to use the image tag by default. We keep
the `manifests_version` parameter, but deprecated it and default it to
the CCM image tag. Users which set the parameter will get a deprecation
warning if they use `manifests_version` with a version that differs from
the configured image tag.

Additionally, by using the best practices `images` config parameter, we
should get automated Renovate version updates for the cloudscale CCM.